potentially more msvc fixes

+++ src/plate.cpp
@@ -143,8 +143,8 @@ void plate::addCrustBySubduction(uint32_t x, uint32_t y, float z, uint32_t t,
         index = _bounds->index(static_cast<uint32_t>(fx), static_cast<uint32_t>(fy));
         if (map[index] > 0)
         {
-            t = (map[index] * age_map[index] + z * t) / (map[index] + z);
-            age_map[index] = static_cast<uint32_t>(t * static_cast<float>(z > 0));
+            t = static_cast<uint32_t>((map[index] * age_map[index] + z * t) / (map[index] + z));
+            age_map[index] = static_cast<uint32_t>(t * static_cast<float>(z > 0.0f));
 
             map[index] += z;
             _mass.incMass(z);
